moving to redux
