+++ components/rhs_header_post/index.jsx
@@ -1,0 +1,21 @@
+// Copyright (c) 2017 Mattermost, Inc. All Rights Reserved.
+// See License.txt for license information.
+
+import {connect} from 'react-redux';
+import {bindActionCreators} from 'redux';
+
+import {showFlaggedPosts, showPinnedPosts, closeRightHandSide} from 'actions/views/rhs';
+
+import RhsHeaderPost from './rhs_header_post.jsx';
+
+function mapDispatchToProps(dispatch) {
+    return {
+        actions: bindActionCreators({
+            showFlaggedPosts,
+            showPinnedPosts,
+            closeRightHandSide
+        }, dispatch)
+    };
+}
+
+export default connect(null, mapDispatchToProps)(RhsHeaderPost);
