+++ reducers/views/rhs.js
@@ -16,6 +16,8 @@ function selectedPostId(state = '', action) {
             return '';
         }
         return state;
+    case ActionTypes.UPDATE_RHS_STATE:
+        return '';
     default:
         return state;
     }
@@ -25,6 +27,8 @@ function selectedPostChannelId(state = '', action) {
     switch (action.type) {
     case ActionTypes.SELECT_POST:
         return action.channelId;
+    case ActionTypes.UPDATE_RHS_STATE:
+        return '';
     default:
         return state;
     }
@@ -36,6 +40,8 @@ function fromSearch(state = '', action) {
         if (action.from_search) {
             return action.from_search;
         }
+        return '';
+    case ActionTypes.UPDATE_RHS_STATE:
         return '';
     default:
         return state;
@@ -49,6 +55,8 @@ function fromFlaggedPosts(state = false, action) {
             return action.from_flagged_posts;
         }
         return false;
+    case ActionTypes.UPDATE_RHS_STATE:
+        return false;
     default:
         return state;
     }
@@ -61,6 +69,8 @@ function fromPinnedPosts(state = false, action) {
             return action.from_pinned_posts;
         }
         return false;
+    case ActionTypes.UPDATE_RHS_STATE:
+        return false;
     default:
         return state;
     }
@@ -70,6 +80,8 @@ function rhsState(state = null, action) {
     switch (action.type) {
     case ActionTypes.UPDATE_RHS_STATE:
         return action.state;
+    case ActionTypes.SELECT_POST:
+        return null;
     default:
         return state;
     }
@@ -81,6 +93,8 @@ function pinnedPostsChannelId(state = null, action) {
         if (action.state === RHSStates.PIN) {
             return action.channelId;
         }
+        return null;
+    case ActionTypes.SELECT_POST:
         return null;
     default:
         return state;
